revert old weather fix and added a better solution

+++ src/game/Weather.cpp
@@ -211,7 +211,7 @@ void Weather::SendFineWeatherUpdateToPlayer(Player* player)
 {
     WorldPacket data(SMSG_WEATHER, (4 + 4 + 4));
 
-    data << (uint32)WEATHER_STATE_FINE << (float)0.0f << uint8(0); // no sound
+    data << (uint32)WEATHER_STATE_FINE << (float)0.0f << uint32(0); // no sound
     player->GetSession()->SendPacket(&data);
 }
